Initialize standalone monitor
Moves the monitor files to its respective folder
