+++ services/monitor/src/types.ts
@@ -1,0 +1,6 @@
+import SourceFetcher from "./source-fetcher";
+
+export type KnownSourceFetchers = {
+  ipfs: SourceFetcher;
+  swarm?: SourceFetcher;
+};
