+++ services/monitor/tsconfig.json
@@ -6,13 +6,13 @@
     "esModuleInterop": true,
     "strict": true,
     "outDir": "dist",
+    "rootDir": "src",
+
     "resolveJsonModule": true,
     "declaration": true,
     "declarationMap": true,
     "module": "commonjs",
-    "paths": {
-      "ethers": ["./node_modules/ethers"]
-    }
   },
-  "include": ["./src", "services/monitor"]
+  "include": ["src/**/*.ts"],
+  "exclude": ["node_modules/**"],
 }
